code/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
scipy==1.11.4
web3==6.11.3
joblib==1.3.2
tensorflow==2.15.0
pandas==2.1.4
scikit-learn==1.3.2
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis for caching and session management
redis==5.0.1
aioredis==2.0.1

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security and Validation
cryptography==41.0.8
email-validator==2.1.0.post1

# Rate limiting
slowapi==0.1.9

# Environment management
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Financial calculations
quantlib==1.32

